moves single files under tests
