+++ alternatives/dotnet/discogs/Program.cs
@@ -43,78 +43,6 @@ namespace discogs
                     await ParseAsync<discogs.Masters.master>(fileName);
                 }
             }
-            else if (Path.GetFileName(fileName) == "label.xml")
-            {
-                DeserializeOneToJson<discogs.Labels.label>(fileName);
-            }
-            else if (Path.GetFileName(fileName) == "release.xml")
-            {
-                DeserializeOneToJson<discogs.Releases.release>(fileName);
-            }
-            else if (Path.GetFileName(fileName) == "master.xml")
-            {
-                DeserializeOneToJson<discogs.Masters.master>(fileName);
-            }
-            else if (Path.GetFileName(fileName) == "artist.xml")
-            {
-                DeserializeOneToJson<discogs.Artists.artist>(fileName);
-            }
-            else if (fileName == "serialize-label")
-            {
-                SerializeLabel();
-            }
-        }
-
-        private static void DeserializeOneToJson<T>(string fileName)
-            where T : IExportToCsv, new()
-        {
-            using var reader = new StreamReader(fileName);
-
-            XmlSerializer _labelXmlSerializer = new XmlSerializer(typeof(T));
-            var obj = (T)_labelXmlSerializer.Deserialize(reader);
-            var jsonOptions = new JsonSerializerOptions
-            {
-                WriteIndented = true,
-            };
-            var objJson = JsonSerializer.Serialize(obj, jsonOptions);
-            Console.WriteLine($@"JSON {typeof(T).Name}:
- {objJson}");
-        }
-
-        private static void SerializeLabel()
-        {
-            var label = new discogs.Labels.label
-            {
-                images = new[] {
-                    new discogs.image { type = "primary", uri="", uri150="", width="132", height="24"}
-                },
-                contactinfo = @"Planet
-                E
-                Communication",
-                parentLabel = new Labels.parentLabel { id = "123", name = "The Parent Label" },
-                data_quality = "Correct",
-                id = "1",
-                name = "Planet E",
-                urls = new[] {
-                    // new discogs.url { TheUrl = "http://planet-e.net" }
-                     "http://planet-e.net"
-                },
-                sublabels = new[] {
-                    new Labels.label { SubId = "2", SubName = "Antidote (4)"}
-                }
-            };
-            var xml = new XmlSerializer(typeof(Labels.label));
-
-            using var writer = new StringWriter();
-            var xmlSettings = new XmlWriterSettings
-            {
-                Indent = true,
-            };
-            using var xmlWriter = XmlWriter.Create(writer);
-            xml.Serialize(writer, label);
-            var labelXml = writer.ToString();
-            Console.WriteLine($@"XML label:
-{labelXml}");
         }
 
         private static async Task ParseAsync<T>(string fileName)
