+++ alternatives/dotnet/tests/Resources/artist.xml
