+++ alternatives/dotnet/tests/Resources/label.xml
