+++ alternatives/dotnet/tests/Resources/master.xml
