+++ alternatives/dotnet/tests/Resources/release.xml
