+++ alternatives/dotnet/tests/tests.csproj
@@ -14,4 +14,15 @@
     <PackageReference Include="coverlet.collector" Version="1.2.0" />
   </ItemGroup>
 
+  <ItemGroup>
+    <EmbeddedResource Include="Resources\artist.xml" />
+    <EmbeddedResource Include="Resources\label.xml" />
+    <EmbeddedResource Include="Resources\master.xml" />
+    <EmbeddedResource Include="Resources\release.xml" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\discogs\discogs.csproj" />
+  </ItemGroup>
+
 </Project>
